Fixed Erdos-Renyi

+++ trunk/src/ComplexNets/GraphGenerator.h
@@ -16,7 +16,7 @@ public:
 
 	Graph generateGraphFromFile(string path, bool directed, bool multigraph);
 	WeightedGraph generateWeightedGraphFromFile(string path, bool directed, bool multigraph);
-	Graph generateErdosRenyiGraph(unsigned int n, unsigned int p);
+	Graph generateErdosRenyiGraph(unsigned int n, float p);
 
 };
